Move write-after-end tests to a dedicated file to avoid concurrent stdio flakiness
